replace RSAEncryptor with EnvelopeEncryptor in prep for KMS
